Update2

+++ mc_pricer.py
@@ -2,6 +2,7 @@ import gbm
 import payoffs as pf
 import numpy as np
 
+
 def discounted_payoffs_call(S0, K, r, sigma, T, n_sims, seed=None):
     ST = gbm.simulate_ST(S0, r, sigma, T, n_sims, seed)
     payoff = pf.call_payoff(ST, K)
@@ -17,10 +18,10 @@ def mc_price_CI(discounted_payoffs):
     n = len(discounted_payoffs)
 
     half_width = 1.96 * std / np.sqrt(n)
-    return mean, mean - half_width, mean + half_width
+    return (float(mean), float(mean - half_width), float(mean + half_width))
 
 def discounted_payoffs_call_antithetic(S0, K, r, sigma, T, n_sims, seed=None):
-   
+    #similiar to original function. Just mirroring generated z scores 
     half = n_sims // 2
     Z = np.random.default_rng(seed).standard_normal(half)
